+++ core/src/main/kotlin/org/pastalab/fray/core/delegates/DelegateSynchronizer.kt
@@ -3,20 +3,25 @@ package org.pastalab.fray.core.delegates
 import org.pastalab.fray.core.RunContext
 import org.pastalab.fray.core.utils.HelperThread
 import org.pastalab.fray.core.utils.Utils
-import org.pastalab.fray.core.utils.Utils.verifyOrReport
 
 class DelegateSynchronizer(val context: RunContext) {
-  var entered = ThreadLocal.withInitial { false }
-  var skipFunctionEntered = ThreadLocal.withInitial { 0 }
-  val stackTrace = ThreadLocal.withInitial { mutableListOf<String>() }
-  val onSkipRecursion = ThreadLocal.withInitial { false }
+  val entered: ThreadLocal<Boolean> = ThreadLocal.withInitial { false }
+
+  val skipPrimitiveEntered: ThreadLocal<Int> = ThreadLocal.withInitial { 0 }
+  val skipPrimitiveStackTrace: ThreadLocal<MutableList<String>> =
+      ThreadLocal.withInitial { mutableListOf() }
+  val onSkipRecursion: ThreadLocal<Boolean> = ThreadLocal.withInitial { false }
+
+  val skipSchedulingEntered: ThreadLocal<Int> = ThreadLocal.withInitial { 0 }
+  val skipSchedulingStackTrace: ThreadLocal<MutableList<String>> =
+      ThreadLocal.withInitial { mutableListOf() }
 
   fun checkEntered(): Boolean {
     if (entered.get()) {
       return true
     }
     entered.set(true)
-    if (skipFunctionEntered.get() > 0) {
+    if (skipPrimitiveEntered.get() > 0) {
       entered.set(false)
       return true
     }
@@ -62,7 +67,7 @@ class DelegateSynchronizer(val context: RunContext) {
       originBlock: (() -> T)
   ): T {
     if (skipDone) {
-      onSkipMethodDone(skipSignature)
+      onSkipPrimitiveDone(skipSignature)
     }
     if (checkEntered()) {
       return originBlock()
@@ -101,13 +106,13 @@ class DelegateSynchronizer(val context: RunContext) {
   ): T {
     if (checkEntered()) {
       if (skipStart) {
-        onSkipMethod(skipSignature)
+        onSkipPrimitive(skipSignature)
       }
       return originBlock()
     }
     val result = frayBlock()
     if (skipStart) {
-      onSkipMethod(skipSignature)
+      onSkipPrimitive(skipSignature)
     }
     entered.set(false)
     result.fold(
@@ -118,22 +123,31 @@ class DelegateSynchronizer(val context: RunContext) {
     )
   }
 
-  fun onSkipMethod(signature: String) {
+  /**
+   * Fray provides two ways to skip instrumented primitive: [onSkipPrimitive] and
+   * [onSkipScheduling]. [onSkipPrimitive] skips the primitives completely, meaning that [Fray] will
+   * not run instrumented code at all. This method should be used if the primitive is modeled
+   * through Fray, and we want to skip its internal implementation (e.g.,
+   * [java.util.concurrent.locks.ReentrantLock]). [onSkipScheduling] on the other hand, still runs
+   * the thread with Fray instrumentation, but it tries to prioritize the thread without scheduling
+   * other threads. This method is useful when a thread enters [<clinit>] or class loaders.
+   */
+  fun onSkipPrimitive(signature: String) {
     if (onSkipRecursion.get()) {
       return
     }
     onSkipRecursion.set(true)
-    stackTrace.get().add(signature)
-    skipFunctionEntered.set(1 + skipFunctionEntered.get())
+    skipPrimitiveStackTrace.get().add(signature)
+    skipPrimitiveEntered.set(1 + skipPrimitiveEntered.get())
 
     if (!context.registeredThreads.containsKey(Thread.currentThread().id)) {
-      stackTrace.get().removeLast()
-      skipFunctionEntered.set(skipFunctionEntered.get() - 1)
+      skipPrimitiveStackTrace.get().removeLast()
+      skipPrimitiveEntered.set(skipPrimitiveEntered.get() - 1)
     }
     onSkipRecursion.set(false)
   }
 
-  fun onSkipMethodDone(signature: String): Boolean {
+  fun onSkipPrimitiveDone(signature: String): Boolean {
     if (onSkipRecursion.get()) {
       return false
     }
@@ -142,13 +156,33 @@ class DelegateSynchronizer(val context: RunContext) {
       if (!context.registeredThreads.containsKey(Thread.currentThread().id)) {
         return false
       }
-      Utils.verifyOrReport(!stackTrace.get().isEmpty())
-      val last = stackTrace.get().removeLast()
+      Utils.verifyOrReport(!skipPrimitiveStackTrace.get().isEmpty())
+      val last = skipPrimitiveStackTrace.get().removeLast()
       Utils.verifyOrReport(last == signature)
-      skipFunctionEntered.set(skipFunctionEntered.get() - 1)
+      skipPrimitiveEntered.set(skipPrimitiveEntered.get() - 1)
       return true
     } finally {
       onSkipRecursion.set(false)
     }
+  }
+
+  fun onSkipScheduling(signature: String) = runInFrayStartNoSkip {
+    skipSchedulingEntered.set(1 + skipSchedulingEntered.get())
+    skipSchedulingStackTrace.get().add(signature)
+    val threadContext = context.registeredThreads[Thread.currentThread().id]!!
+    context.prioritizedThreads.add(threadContext)
+    return@runInFrayStartNoSkip Result.success(Unit)
+  }
+
+  fun onSkipSchedulingDone(signature: String) = runInFrayDoneNoSkip {
+    Utils.verifyOrReport(!skipSchedulingStackTrace.get().isEmpty())
+    val last = skipSchedulingStackTrace.get().removeLast()
+    Utils.verifyOrReport(last == signature)
+    skipSchedulingEntered.set(skipSchedulingEntered.get() - 1)
+    if (skipSchedulingEntered.get() == 0) {
+      val threadContext = context.registeredThreads[Thread.currentThread().id]!!
+      context.prioritizedThreads.remove(threadContext)
+    }
+    return@runInFrayDoneNoSkip Result.success(Unit)
   }
 }