+++ instrumentation/base/src/main/kotlin/org/pastalab/fray/instrumentation/base/visitors/ForkJoinPoolInstrumenter.kt
@@ -4,7 +4,6 @@ import java.util.concurrent.ForkJoinPool
 import java.util.concurrent.ForkJoinTask
 import org.objectweb.asm.ClassVisitor
 import org.objectweb.asm.MethodVisitor
-import org.objectweb.asm.Opcodes.ASM9
 import org.objectweb.asm.Type
 import org.objectweb.asm.commons.AdviceAdapter
 import org.pastalab.fray.runtime.Runtime