+++ instrumentation/base/src/main/kotlin/org/pastalab/fray/instrumentation/base/visitors/SkipPrimitiveInstrumenter.kt
@@ -1,0 +1,11 @@
+package org.pastalab.fray.instrumentation.base.visitors
+
+import org.objectweb.asm.ClassVisitor
+
+class SkipPrimitiveInstrumenter(cv: ClassVisitor) :
+    SkipMethodInstrumenter(
+        cv,
+        org.pastalab.fray.runtime.Runtime::onSkipPrimitive,
+        org.pastalab.fray.runtime.Runtime::onSkipPrimitiveDone,
+        "com.intellij.rt.debugger",
+        "org.gradle.internal")